Please Review This
This is part of my weapon handling system for  new shooter . The player
has 2 slots to store weapons, trying to pick up a third weapon will swap
the current weapon being held with the new. When a weapon is not
occupying the slot then an empty gun object is set in place.